build: make forward declaration for acl structs
